+++ ksyun/client/ebs/v20160304/models.py
@@ -47,7 +47,7 @@ ESSD_PL3取值范围：【1761，32768】，步长：1GB
         :param Tag: 标签属性
         :type PathPrefix: Filter
         :param ProvisionedIops: ESSD_AutoPL云盘的预配置值,仅AutoPL类型云盘可用
-        :type PathPrefix: Int
+        :type PathPrefix: Long
         """
         self.VolumeName = None
         self.VolumeType = None
